kgq10h: Extract shared base class for encrypted Wouxun protocol
Move common methods (encrypt, decrypt, checksum, download, tone
encode/decode, duplex/offset decode, name decode) into WouxunKGBase
in wouxun_kg_common.py. This shared base class serves the Wouxun
encrypted serial protocol family (KG-Q10H, KG-935G). The existing
wouxun_common.py covers the older unencrypted protocol (KG-UV6 etc.).

KGQ10HRadio now inherits from WouxunKGBase instead of directly from
CloneModeRadio + ExperimentalRadio.
